Tools/FaceID/requirements.txt
facenet-pytorch>=2.5.3
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
